chore(core): move test_assert.h and test_color.h to common
