+++ core/tests/unit/ldml/test_kmx_plus.cpp
@@ -1,4 +1,4 @@
-#include "../test_assert.h"
+#include <test_assert.h>
 #include "kmx/kmx_plus.h"
 #include "kmx/kmx_xstring.h"
 #include "../../../src/ldml/ldml_vkeys.hpp"
